Faster and maybe more optimized
insted of finding paths like BFS it tries to find two path one for target and one for source to Kevin Bacon. Which takes only Six steps at miximum O(6) to find for each Twelve steps total and the rest of the program. Then it mergs both the stacks at Kevin backon (Node).